Latest working build
Sound still doesn't fucking work. Kill me now.
Commented code to point out what I have worked on (not a lot, as it turns out). Many things I've made have been moved and I can't find it. was it deleted??
Modified movePlayer() in levellogic to make the entity with isPlayer=true move instead of the first entity in the list of entities.

The level creator now retrieves entitydescription. We need to modify the level creator further down the line to account for entitytype becoming redundant.